// pwa/src/generated/models/messageType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wetty Chat API
 * Real-time chat application backend API supporting groups, messaging, threads, stickers, invites, and push notifications.
 * OpenAPI spec version: 0.1.0
 */

export enum MessageType {
  text = 'text',
  audio = 'audio',
  file = 'file',
  sticker = 'sticker',
  invite = 'invite',
  system = 'system',
}